refactor: 优化name_match结构
