+++ src/policy/pkg_cfg.rs
@@ -1,5 +1,5 @@
 use super::{
-    name_match::{policy_ue, policy_unity},
+    name_match::policies::{policy_ue, policy_unity},
     usage::{
         usage_top1::policies::{policy_cocos, policy_second, policy_top1, policy_ue5},
         usage_top2::policy_top2,
